Unify test files
